% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearcut.R
\name{earcut}
\alias{earcut}
\alias{earcut}
\alias{earcut.default}
\title{Ear cut algorithm}
\usage{
earcut(x, holes = NULL, ...)

\method{earcut}{default}(x, holes = NULL, ...)
}
\arguments{
\item{x}{object to triangulate, any convertible to silicate PATH}

\item{holes}{1-based index indicating where hole/s begin}

\item{...}{arguments passed to methods}
}
\value{
index triplets of triangles from vertex pool
}
\description{
Ear cutting for polygons, a method for constrained triangulation.
}
