% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearcut-package.R
\name{earcut}
\alias{earcut}
\alias{earcut}
\alias{earcut.default}
\alias{earcut}
\alias{earcut.PATH}
\title{Ear cut algorithm}
\usage{
earcut(x, holes = NULL...)

\method{earcut}{default}(x, holes = NULL, ...)

\method{earcut}{PATH}(x, ...)
}
\arguments{
\item{x}{object to triangulate, any convertible to silicate PATH}

\item{...}{arguments passed to methods}
}
\value{
index triplets of triangles from vertex pool
}
\description{
Ear cutting for polygons, a method for constrained triangulation.
}
