% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearcut-package.R
\docType{package}
\name{rearcut-package}
\alias{rearcut}
\alias{rearcut-package}
\alias{rearcut-package}
\title{rearcut: Constrained Triangulation of Polygons by 'Ear Clipping (Cutting)'}
\description{
Constrained triangulation of polygons by direct port of the 'Mapbox' library 'earcut'.

rearcut
R package wrapping the Mapbox library earcut.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
