% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rearcut-package.R
\docType{data}
\name{rearcut}
\alias{rearcut}
\alias{rearcut}
\title{Polygon triangulation}
\usage{
rearcut(x, y)
}
\arguments{
\item{x}{numeric vector x coordinates}

\item{y}{numeric vector y coordinates}
}
\description{
Constrained polygon triangulation by ear cut algorithm,
also known as ear clipping.

rearcut
R port of Mabbox earcut.
}
\examples{
X <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69, 0)
Y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0, 0)
ind <- rearcut(X, Y)  ## should be longer than 0 :)
if (length(ind) < 3) {
 ## this is rgl's solution, it's not necessarily deterministic
 ind <- c(8L, 2L, 1L, 8L, 5L, 2L, 8L, 7L, 5L, 5L, 3L, 2L, 5L,
              4L, 3L, 7L, 6L, 5L) - 1
}
plot(X, Y)
apply(matrix(ind, nrow = 3), 2, function(x) polygon(cbind(X, Y)[x + 1, ]))
}
